Built out open function in Database Class; Created DatabaseException class.
The open function still has a little work that needs to be done at this
point. The Database Exception isn't very unique from Exception just yet,
but will need to figure out what might be some useful functions or
information in this exception that might not otherwise be needed for a
general exception.